+++ config/agents/app/kustomization.yaml
@@ -1,0 +1,8 @@
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+resources:
+- agentapp.yaml
+- role.yaml
+- role_binding.yaml
+- service_account.yaml
+
