+++ config/agents/app/role_binding.yaml
@@ -1,0 +1,18 @@
+apiVersion: rbac.authorization.k8s.io/v1
+kind: RoleBinding
+metadata:
+  labels:
+    app.kubernetes.io/name: rolebinding
+    app.kubernetes.io/instance: agentapp-rolebinding
+    app.kubernetes.io/component: rbac
+    app.kubernetes.io/created-by: primaza
+    app.kubernetes.io/part-of: primaza
+    app.kubernetes.io/managed-by: kustomize
+  name: manager-rolebinding
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: Role
+  name: agentapp-role
+subjects:
+- kind: ServiceAccount
+  name: controller-agentapp
