+++ config/crd/overlays/agentapp/kustomization.yaml
@@ -1,0 +1,3 @@
+resources:
+- ../../bases/primaza.io_servicebindings.yaml
+
