+++ make/agents/app/agentapp.mk
@@ -1,0 +1,3 @@
+AGENTAPP_DOCKERFILE=$(PROJECT_DIR)/deploy/agents/app/Dockerfile
+
+IMG ?= agentapp:latest
